app/requirements.txt
fastapi
uvicorn[standard]
pyjwt
redis
fastapi-limiter